+++ androidx-kt/src/main/java/me/panpf/androidxkt/graphics/Color.kt
@@ -16,6 +16,7 @@
 
 package me.panpf.androidxkt.graphics
 
+import android.animation.ArgbEvaluator
 import android.graphics.Color
 import android.graphics.ColorMatrix
 import android.graphics.ColorMatrixColorFilter
@@ -548,4 +549,26 @@ fun Int.makeMatrixColorFilter(): ColorMatrixColorFilter {
     val matrix = ColorMatrix()
     matrix.set(src)
     return ColorMatrixColorFilter(src)
+}
+
+
+/**
+ * This function returns the calculated in-between value for a color
+ * given integers that represent the start and end values in the four
+ * bytes of the 32-bit int. Each channel is separately linearly interpolated
+ * and the resulting calculated values are recombined into the return value.
+ *
+ * @receiver startValue A 32-bit int value representing colors in the
+ * @param endValue A 32-bit int value representing colors in the
+ * @param fraction The fraction from the starting to the ending values
+ * separate bytes of the parameter
+ * separate bytes of the parameter
+ * @return A value that is calculated to be the linearly interpolated
+ * result, derived by separating the start and end values into separate
+ * color channels and interpolating each one separately, recombining the
+ * resulting values in the same way.
+ */
+@ColorInt
+fun Int.colorArgbEvaluate(@ColorInt endValue: Int, @FloatRange(from = 0.0, to = 1.0) fraction: Float): Int {
+    return ArgbEvaluator().evaluate(fraction, this, endValue) as Int
 }