+++ androidx/src/main/java/me/panpf/androidx/graphics/Colorx.java
@@ -16,6 +16,7 @@
 
 package me.panpf.androidx.graphics;
 
+import android.animation.ArgbEvaluator;
 import android.graphics.Color;
 import android.graphics.ColorMatrix;
 import android.graphics.ColorMatrixColorFilter;
@@ -536,5 +537,26 @@ public class Colorx {
         ColorMatrix matrix = new ColorMatrix();
         matrix.set(src);
         return new ColorMatrixColorFilter(src);
+    }
+
+    /**
+     * This function returns the calculated in-between value for a color
+     * given integers that represent the start and end values in the four
+     * bytes of the 32-bit int. Each channel is separately linearly interpolated
+     * and the resulting calculated values are recombined into the return value.
+     *
+     * @param startValue A 32-bit int value representing colors in the
+     * separate bytes of the parameter
+     * @param endValue A 32-bit int value representing colors in the
+     * @param fraction The fraction from the starting to the ending values
+     * separate bytes of the parameter
+     * @return A value that is calculated to be the linearly interpolated
+     * result, derived by separating the start and end values into separate
+     * color channels and interpolating each one separately, recombining the
+     * resulting values in the same way.
+     */
+    @ColorInt
+    public static int argbEvaluate(@ColorInt int startValue, @ColorInt int endValue, @FloatRange(from = 0, to = 1) float fraction) {
+        return (int) new ArgbEvaluator().evaluate(fraction, startValue, endValue);
     }
 }