+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * 增加 Text 增加 keywordMadeColorBySpannable() 方法用于将文件中关键词部分改变颜色
 * 重构 Color
 * 升级 Javax 依赖
+* Color 增加 argbEvaluate 方法
 
 ## v0.1
 * 初次发布